requirements.txt
streamlit
sentence-transformers
scikit-learn
numpy
pandas
pdfplumber
python-docx
rapidfuzz
faiss-cpu
plotly
python-multipart
python-dotenv
joblib
google-generative-ai
requests
reportlab
kaleido
streamlit-lottie
